requirements.txt
graphtsetlinmachine
numpy
numba
pycuda
scikit-learn
pydantic
ruamel.yaml
pydantic
pandas
sympy
argparse